requirements.txt
#pip install -r requirements.txt

hx711
mysql-connector-python
flask